refactor: pr 수정요청사항 적용
