+++ frontend/techpick/src/app/style.css.ts
@@ -1,6 +1,6 @@
 import { style } from '@vanilla-extract/css';
 
-export const rootWrapper = style({
+export const rootLayout = style({
   width: '100%',
   height: '100vh',
 });
