+++ frontend/techpick/src/shared/themes/themeDark.css.ts
@@ -1,0 +1,12 @@
+import { createTheme } from '@vanilla-extract/css';
+import { themeContract } from '@/shared/themes/themeContract.css';
+
+export const darkTheme = createTheme(themeContract, {
+  color: {
+    primary: '#00306e',
+    secondary: '#95004d',
+    border: '#eaeaea',
+    background: '#000',
+    font: '#fff',
+  },
+});
