+++ frontend/techpick/src/widgets/RightSection/rightSection.css.ts
@@ -1,0 +1,6 @@
+import { style } from '@vanilla-extract/css';
+
+export const rightSection = style({
+  width: '0%',
+  height: '100%',
+});
