+++ frontend/techpick/src/widgets/index.ts
@@ -1,0 +1,4 @@
+export { default as MainView } from './MainView/MainView';
+export { default as LeftSection } from './LeftSection/LeftSection';
+export { default as MiddleSection } from './MiddleSection/MiddleSection';
+export { default as RightSection } from './RightSection/RightSection';
